+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ## unreleased
 * `Card`:
-  * Update `CardRequest` to require `ThreeDSecureRequest`
+  * Remove `ThreeDSecureRequest` from `CardRequest` and create URLs internally 
+  * Update `CardRequest` to optionally pass `sca`
 
 ## 0.0.3 (2022-11-03)
 * PayPalNativeCheckout